feat(workflow): support limit count action (#7771)
