+++ apisix/plugins/workflow.lua
@@ -14,10 +14,10 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 --
-local core       = require("apisix.core")
-local expr       = require("resty.expr.v1")
-local ipairs     = ipairs
-local tonumber   = tonumber
+local core        = require("apisix.core")
+local limit_count = require("apisix.plugins.limit-count.init")
+local expr        = require("resty.expr.v1")
+local ipairs      = ipairs
 
 local schema = {
     type = "object",
@@ -71,17 +71,34 @@ local return_schema = {
 }
 
 
+local function check_return_schema(conf)
+    local ok, err = core.schema.check(return_schema, conf)
+    if not ok then
+        return false, err
+    end
+    return true
+end
+
+
 local function exit(conf)
-    local code = tonumber(conf.code)
-    return code, {error_msg = "rejected by workflow"}
+    return conf.code, {error_msg = "rejected by workflow"}
+end
+
+
+local function rate_limit(conf, ctx)
+    return limit_count.rate_limit(conf, ctx)
 end
 
 
 local support_action = {
     ["return"] = {
-        handler  = exit,
-        schema   = return_schema,
+        handler        = exit,
+        check_schema   = check_return_schema,
     },
+    ["limit-count"] = {
+        handler        = rate_limit,
+        check_schema   = limit_count.check_schema,
+    }
 }
 
 
@@ -91,7 +108,7 @@ function _M.check_schema(conf)
         return false, err
     end
 
-    for _, rule in ipairs(conf.rules) do
+    for idx, rule in ipairs(conf.rules) do
         local ok, err = expr.new(rule.case)
         if not ok then
             return false, "failed to validate the 'case' expression: " .. err
@@ -104,7 +121,9 @@ function _M.check_schema(conf)
                 return false, "unsupported action: " .. action[1]
             end
 
-            local ok, err = core.schema.check(support_action[action[1]].schema, action[2])
+            -- use the action's idx as an identifier to isolate between confs
+            action[2]["_vid"] = idx
+            local ok, err = support_action[action[1]].check_schema(action[2], plugin_name)
             if not ok then
                 return false, "failed to validate the '" .. action[1] .. "' action: " .. err
             end
@@ -123,7 +142,7 @@ function _M.access(conf, ctx)
         if match_result then
             -- only one action is currently supported
             local action = rule.actions[1]
-            return support_action[action[1]].handler(action[2])
+            return support_action[action[1]].handler(action[2], ctx)
         end
     end
 end
